requirements.txt
streamlit>=1.34
requests>=2.31.0
beautifulsoup4>=4.12.3
youtube-transcript-api>=0.6.2
python-dotenv>=1.0.1
chromadb>=0.5.5
python-docx>=0.8.11

